Moved setup controller methods out of settings controller
